// src/models/IoK8sApiAdmissionregistrationV1beta1ValidatingAdmissionPolicyBinding/types.ts
import { IoK8sApiAdmissionregistrationV1beta1ValidatingAdmissionPolicyBindingSpecModelOptions as IoK8sApiAdmissionregistrationV1beta1ValidatingAdmissionPolicyBindingSpec } from '../IoK8sApiAdmissionregistrationV1beta1ValidatingAdmissionPolicyBindingSpec';
import { IoK8sApimachineryPkgApisMetaV1ObjectMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ObjectMeta } from '../IoK8sApimachineryPkgApisMetaV1ObjectMeta';

export interface IoK8sApiAdmissionregistrationV1beta1ValidatingAdmissionPolicyBindingModelOptions {
  apiVersion?: string;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ObjectMeta;
  spec?: IoK8sApiAdmissionregistrationV1beta1ValidatingAdmissionPolicyBindingSpec;
}